Add isDefault flag to ruleGroup (#8936)

+++ pkg/handler/v2/rulegroup/rulegroup.go
@@ -109,6 +109,9 @@ func UpdateEndpoint(userInfoGetter provider.UserInfoGetter, projectProvider prov
 		if err != nil {
 			return nil, common.KubernetesErrorToHTTPError(err)
 		}
+		if currentRuleGroup.Spec.IsDefault {
+			return nil, utilerrors.NewBadRequest("only Admin can update default rule group")
+		}
 		ruleGroup, err := convertAPIToInternalRuleGroup(c, &req.Body, req.RuleGroupID)
 		if err != nil {
 			return nil, common.KubernetesErrorToHTTPError(err)
@@ -130,6 +133,13 @@ func DeleteEndpoint(userInfoGetter provider.UserInfoGetter, projectProvider prov
 		c, err := handlercommon.GetCluster(ctx, projectProvider, privilegedProjectProvider, userInfoGetter, req.ProjectID, req.ClusterID, nil)
 		if err != nil {
 			return nil, err
+		}
+		currentRuleGroup, err := getRuleGroup(ctx, userInfoGetter, c, req.ProjectID, req.RuleGroupID)
+		if err != nil {
+			return nil, common.KubernetesErrorToHTTPError(err)
+		}
+		if currentRuleGroup.Spec.IsDefault {
+			return nil, utilerrors.NewBadRequest("only Admin can delete default rule group")
 		}
 		if err = deleteRuleGroup(ctx, userInfoGetter, c, req.ProjectID, req.RuleGroupID); err != nil {
 			return nil, common.KubernetesErrorToHTTPError(err)
@@ -229,7 +239,8 @@ func convertAPIToInternalRuleGroup(cluster *kubermaticv1.Cluster, ruleGroup *api
 				APIVersion:      cluster.APIVersion,
 				ResourceVersion: cluster.ResourceVersion,
 			},
-			Data: ruleGroup.Data,
+			Data:      ruleGroup.Data,
+			IsDefault: ruleGroup.IsDefault,
 		},
 	}
 	return internalRuleGroup, nil
@@ -245,9 +256,10 @@ func convertInternalToAPIRuleGroups(ruleGroups []*kubermaticv1.RuleGroup) []*api
 
 func convertInternalToAPIRuleGroup(ruleGroup *kubermaticv1.RuleGroup) *apiv2.RuleGroup {
 	return &apiv2.RuleGroup{
-		Name: ruleGroup.ObjectMeta.Name,
-		Data: ruleGroup.Spec.Data,
-		Type: ruleGroup.Spec.RuleGroupType,
+		Name:      ruleGroup.ObjectMeta.Name,
+		Data:      ruleGroup.Spec.Data,
+		Type:      ruleGroup.Spec.RuleGroupType,
+		IsDefault: ruleGroup.Spec.IsDefault,
 	}
 }
